moved login button to auth
